Move routing helper methods to more appropriate package
